scripts/requirements.txt
httpx==0.27.0
anthropic>=0.50.0
python-dotenv==1.0.1
pytest==8.3.2
respx==0.21.1
pytest-asyncio==0.24.0
